pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "show-scribe"
version = "0.1.0"
description = "Audio analysis and review pipeline for Show-Scribe."
readme = "README.md"
authors = [{ name = "Show-Scribe Team" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.7",
    "PyYAML>=6.0",
    "pydantic>=2.5",
    "numpy>=1.25",
    "pandas>=2.1",
    "scipy>=1.11",
    "soundfile>=0.12",
    "librosa>=0.10",
    "resampy>=0.4.2",
    "tqdm>=4.66",
    "faster-whisper>=1.0.0",
    "torch>=2.1",
    "torchaudio>=2.1",
    "demucs>=4.0.0",
    "resemble-enhance>=0.0.1",
    "clearvoice>=0.1.2",
    "pyannote.audio>=3.1.1",
    "resemblyzer>=0.1.1",
    "requests>=2.31",
    "streamlit>=1.29",
    "jsonschema>=4.19",
    "huggingface-hub>=0.20",
]

[project.optional-dependencies]
dev = [
    "black>=24.2",
    "ruff>=0.5.0",
    "mypy>=1.8",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pre-commit>=3.5",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31",
]
preprocessing = [
    "audio-separator>=0.17.0",
    "resemble-enhance>=0.0.1",
    "clearvoice>=0.1.2",
]

[project.scripts]
"show-scribe" = "show_scribe.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["show_scribe*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = "src/|tests/"
extend-exclude = """
(
  src/show_scribe/config/schema.json
  | src/show_scribe/storage/schema.sql
  | tests/fixtures
)
"""

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests.",
    "slow: marks tests as slow running.",
]

[tool.coverage.run]
branch = true
source = ["src/show_scribe"]
omit = ["*/__init__.py"]

[tool.coverage.report]
fail_under = 80
skip_covered = true
show_missing = true
